feat(docs): add some proper generated docs with Typedoc
- Clean up a lot of types
  - Rename a heck ton of types in the back
  - Export a lot of missing types that could be  useful
