+++ package.json
@@ -32,6 +32,7 @@
     "np": "^7.6.1",
     "prettier": "2.7.1",
     "simple-git-hooks": "^2.8.0",
+    "typedoc": "0.23.24",
     "typescript": "^4.6.3"
   },
   "scripts": {
@@ -44,7 +45,8 @@
     "release": "yarn build && changeset publish",
     "test": "vitest run",
     "test:ui": "vitest --ui",
-    "ts": "cd packages/react-three-rapier && yarn ts"
+    "ts": "cd packages/react-three-rapier && yarn ts",
+    "r3r-docs": "typedoc --out ./packages/react-three-rapier/docs ./packages/react-three-rapier/src/index.ts --excludeInternal"
   },
   "simple-git-hooks": {
     "commit-msg": "npx --no -- commitlint --edit ${1}",
